[UPDATED] Refactored station related code into its own package.
