+++ train-departures-android/src/main/java/uk/co/baconi/pka/td/SearchCriteriaAdapter.kt
@@ -9,7 +9,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ArrayAdapter
 import android.widget.TextView
-import uk.co.baconi.pka.common.StationCode
+import uk.co.baconi.pka.common.stations.StationCode
 
 class SearchCriteriaAdapter(
     context: Context,
